Formatting docs to ApiGen format, added minor documentation, added todos
ApiGen likes their docblocks to look as follows:
@param [type] [desc]
All the documentation was set up as:
@param [type] [varname] [desc]
Changed everything to be in first type, as that's what we're using.
ApiGen is smart enough to determine the varname from the function
itself.

Also added a bunch of `<var>varname</var>` to the **Argument**
descriptors. I realized that there's no good way to list the arguments
next to constants, but we are able to write raw html into the docblocks.
This makes the API docs significantly more readable.

Also added some text to the eventinterface header.

Also added a few todos where either (1) there's no docs or (2) the docs
expose ehough_* methods which we don't want people to see.